# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TALABAT Multi-Restaurant SaaS Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /restaurants/{id}/categories:
    get:
      operationId: listCategories
      tags: [categories]
      summary: List categories for a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Category"

    post:
      operationId: createCategory
      tags: [categories]
      summary: Create category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryInput"
      responses:
        "201":
          description: Category created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"
  /categories/{id}:
    patch:
      operationId: updateCategory
      tags: [categories]
      summary: Update category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CategoryPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Category"

    delete:
      operationId: deleteCategory
      tags: [categories]
      summary: Delete category
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted

  # ── Products ────────────────────────────────────────────────────────────────
...
  /restaurants/{id}/drivers:
    get:
      operationId: listDrivers
      tags: [drivers]
      summary: List drivers for a restaurant
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Drivers
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"

    post:
      operationId: createDriver
      tags: [drivers]
      summary: Create driver
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Driver created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    patch:
      operationId: updateDriver
      tags: [drivers]
      summary: Update driver
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverPatch"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"

    delete:
      operationId: deleteDriver
      tags: [drivers]
      summary: Delete driver
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "204":
          description: Deleted

  # ── Orders ──────────────────────────────────────────────────────────────────
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List all orders (admin)
      parameters:
        - name: restaurantId
          in: query
          schema: { type: number }
        - name: status
          in: query
          schema: { type: string }
        - name: driverId
          in: query
          schema: { type: number }
        - name: dateFrom
          in: query
          schema: { type: string, format: date }
        - name: dateTo
          in: query
          schema: { type: string, format: date }
        - name: search
          in: query
          schema: { type: string }
        - name: page
          in: query
          schema: { type: number, default: 1 }
        - name: limit
          in: query
          schema: { type: number, default: 20 }
      responses:
        "200":
          description: Orders list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
  /orders/{id}:
    get:
      operationId: getOrder
      tags: [orders]
      summary: Get order by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
        "404":
          description: Not found
  /orders/{id}/status:
    patch:
      operationId: updateOrderStatus
      tags: [orders]
      summary: Update order status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrderStatusInput"
      responses:
        "200":
          description: Updated order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"
  /orders/{id}/assign-driver:
    post:
      operationId: assignDriver
      tags: [orders]
      summary: Assign driver to order
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverAssignment"
      responses:
        "200":
          description: Driver assigned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderDetail"

  # ── Public storefront ────────────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List admin notifications
      parameters:
        - name: unreadOnly
          in: query
          schema: { type: boolean }
        - name: page
          in: query
          schema: { type: number, default: 1 }
        - name: limit
          in: query
          schema: { type: number, default: 20 }
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationListResponse"
  /notifications/{id}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: number }
      responses:
        "200":
          description: Marked read
...
components:
  schemas:
...
    Category:
      type: object
      required: [id, restaurantId, name, sortOrder, isAvailable]
      properties:
        id: { type: number }
        restaurantId: { type: number }
        name: { type: string }
        nameAr: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        sortOrder: { type: number }
        isAvailable: { type: boolean }
    CategoryInput:
      type: object
      required: [name]
      properties:
        name: { type: string, minLength: 1 }
        nameAr: { type: string }
        imageUrl: { type: string }
        sortOrder: { type: number }
        isAvailable: { type: boolean }
    CategoryPatch:
      type: object
      properties:
        name: { type: string }
        nameAr: { type: ["string", "null"] }
        imageUrl: { type: ["string", "null"] }
        sortOrder: { type: number }
        isAvailable: { type: boolean }

    # ── Product ─────────────────────────────────────────────────────────────────
...
    Driver:
      type: object
      required: [id, restaurantId, name, phone, isActive, totalDeliveries]
      properties:
        id: { type: number }
        restaurantId: { type: number }
        name: { type: string }
        phone: { type: string }
        vehicleType: { type: ["string", "null"] }
        vehiclePlate: { type: ["string", "null"] }
        isActive: { type: boolean }
        totalDeliveries: { type: number }
        createdAt: { type: string, format: date-time }
    DriverInput:
      type: object
      required: [name, phone]
      properties:
        name: { type: string, minLength: 1 }
        phone: { type: string }
        vehicleType: { type: string }
        vehiclePlate: { type: string }
        isActive: { type: boolean }
    DriverPatch:
      type: object
      properties:
        name: { type: string }
        phone: { type: string }
        vehicleType: { type: ["string", "null"] }
        vehiclePlate: { type: ["string", "null"] }
        isActive: { type: boolean }
    DriverAssignment:
      type: object
      required: [driverId]
      properties:
        driverId: { type: number }

    # ── Order ───────────────────────────────────────────────────────────────────
    Order:
      type: object
      required: [id, restaurantId, customerName, customerPhone, totalAmount, status, createdAt]
      properties:
        id: { type: number }
        restaurantId: { type: number }
        restaurantName: { type: string }
        customerName: { type: string }
        customerPhone: { type: string }
        notes: { type: ["string", "null"] }
        latitude: { type: number }
        longitude: { type: number }
        mapsUrl: { type: string }
        totalAmount: { type: number }
        status:
          type: string
          enum: [NEW, ACCEPTED, PREPARING, READY, OUT_FOR_DELIVERY, DELIVERED, CANCELLED]
        driverId: { type: ["number", "null"] }
        driverName: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
    OrderDetail:
      type: object
      required: [id, restaurantId, customerName, customerPhone, totalAmount, status, items, statusHistory, createdAt]
      properties:
        id: { type: number }
        restaurantId: { type: number }
        restaurantName: { type: string }
        customerName: { type: string }
        customerPhone: { type: string }
        notes: { type: ["string", "null"] }
        latitude: { type: number }
        longitude: { type: number }
        mapsUrl: { type: string }
        totalAmount: { type: number }
        status:
          type: string
          enum: [NEW, ACCEPTED, PREPARING, READY, OUT_FOR_DELIVERY, DELIVERED, CANCELLED]
        driverId: { type: ["number", "null"] }
        driverName: { type: ["string", "null"] }
        driverPhone: { type: ["string", "null"] }
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        statusHistory:
          type: array
          items:
            $ref: "#/components/schemas/StatusHistoryEntry"
        createdAt: { type: string, format: date-time }
    OrderItem:
      type: object
      required: [id, productId, productName, quantity, unitPrice, subtotal]
      properties:
        id: { type: number }
        productId: { type: number }
        productName: { type: string }
        quantity: { type: number }
        unitPrice: { type: number }
        subtotal: { type: number }
        selectedAddons:
          type: array
          items:
            $ref: "#/components/schemas/SelectedAddon"
    SelectedAddon:
      type: object
      required: [addonId, addonName, price]
      properties:
        addonId: { type: number }
        addonName: { type: string }
        price: { type: number }
    StatusHistoryEntry:
      type: object
      required: [status, changedAt]
      properties:
        status: { type: string }
        changedAt: { type: string, format: date-time }
        note: { type: ["string", "null"] }
    OrderListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total: { type: number }
        page: { type: number }
        limit: { type: number }
    OrderStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [NEW, ACCEPTED, PREPARING, READY, OUT_FOR_DELIVERY, DELIVERED, CANCELLED]
        note: { type: string }
...
    Notification:
      type: object
      required: [id, type, message, isRead, createdAt]
      properties:
        id: { type: number }
        type: { type: string }
        message: { type: string }
        relatedId: { type: ["number", "null"] }
        relatedType: { type: ["string", "null"] }
        isRead: { type: boolean }
        createdAt: { type: string, format: date-time }
    NotificationListResponse:
      type: object
      required: [data, total, unreadCount]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Notification"
        total: { type: number }
        unreadCount: { type: number }

    # ── Settings ─────────────────────────────────────────────────────────────────
...